lots or reorg. Now using analysis package for more felxibility like future use of ssa
